Feature/#11 use query pages

+++ src/api.js
@@ -1,9 +1,12 @@
+/* eslint-disable prefer-destructuring */
 /* eslint-disable no-alert */
 /* eslint-disable object-shorthand */
 import axios from 'axios';
 import { useQuery } from 'react-query';
 import fileDownload from 'js-file-download';
 import { url } from './url';
+
+let token = JSON.parse(localStorage.getItem('user'))?.tokens?.accessToken;
 
 async function getAnnouncements(page, size) {
     const { data } = await axios.get(`${url}/announcement?page=${page}&size=${size}`);
@@ -90,28 +93,30 @@ export function fetchCalendarData() {
     return fetch(`${url}/calendar`).then((response) => response.json());
 }
 
-export async function addToDo(data, setToggleAddBox) {
+export async function addToDo(data, setToggleAddBox, setChangeTodo) {
     if (data.title === '') {
         alert('제목을 입력하세요.');
     } else if (data.startDate === '') {
         alert('날짜를 선택하세요.');
-    } else {
+    } else if (token) {
         await axios({
             method: 'post',
             url: `${url}/calendar`,
             data,
             headers: {
-                Authorization:
-                    'Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJybGF0anJ1ZDEyMzIiLCJ1c2VySWQiOiJybGF0anJ1ZDEyMzIiLCJlbWFpbCI6InJsYXRqcnVkMTExQGdtYWlsLmNvbSIsIm5hbWUiOiLquYDshJzqsr0iLCJyb2xlIjoi7ZqM7JuQIiwiaWF0IjoxNjQ5NDA3MTgyLCJleHAiOjE2NDk0OTM1ODJ9.Cw4UJWRodHiDhOeaN-8pg3Bboa8dppDKzVoaWgaL1VY',
+                Authorization: `Bearer ${token}`,
             },
         })
             .then(() => {
                 alert('일정이 추가되었습니다.');
                 setToggleAddBox(false);
+                setChangeTodo((prev) => prev + 1);
             })
             .catch((error) => {
                 console.log(error.response);
             });
+    } else {
+        alert('권한이 없습니다.');
     }
 }
 
@@ -121,8 +126,7 @@ export async function fetchToDoList(date, setToDoList) {
         method: 'get',
         url: `${url}/calendar?year=${year}`,
         headers: {
-            Authorization:
-                'Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJybGF0anJ1ZDEyMzIiLCJ1c2VySWQiOiJybGF0anJ1ZDEyMzIiLCJlbWFpbCI6InJsYXRqcnVkMTExQGdtYWlsLmNvbSIsIm5hbWUiOiLquYDshJzqsr0iLCJyb2xlIjoi7ZqM7JuQIiwiaWF0IjoxNjQ5NDA3MTgyLCJleHAiOjE2NDk0OTM1ODJ9.Cw4UJWRodHiDhOeaN-8pg3Bboa8dppDKzVoaWgaL1VY',
+            Authorization: `Bearer ${token}`,
         },
     })
         .then((response) => {
@@ -146,20 +150,24 @@ export async function fetchToDoList(date, setToDoList) {
         });
 }
 
-export async function deleteToDo(e) {
-    if (window.confirm(`${e.target.innerHTML}를 삭제하시겠습니까?`)) {
-        await axios({
-            method: 'delete',
-            url: `${url}/calendar/${e.target.id}`,
-            headers: {
-                Authorization:
-                    'Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJybGF0anJ1ZDEyMzIiLCJ1c2VySWQiOiJybGF0anJ1ZDEyMzIiLCJlbWFpbCI6InJsYXRqcnVkMTExQGdtYWlsLmNvbSIsIm5hbWUiOiLquYDshJzqsr0iLCJyb2xlIjoi7ZqM7JuQIiwiaWF0IjoxNjQ5NDA3MTgyLCJleHAiOjE2NDk0OTM1ODJ9.Cw4UJWRodHiDhOeaN-8pg3Bboa8dppDKzVoaWgaL1VY',
-            },
-        })
-            .then(alert('삭제되었습니다.'))
-            .catch((error) => {
-                console.log(error);
-            });
+export async function deleteToDo(e, setChangeTodo) {
+    if (token) {
+        if (window.confirm(`${e.target.innerHTML}를 삭제하시겠습니까?`)) {
+            await axios({
+                method: 'delete',
+                url: `${url}/calendar/${e.target.id}`,
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+            })
+                .then(() => {
+                    alert('삭제되었습니다.');
+                    setChangeTodo((prev) => prev + 1);
+                })
+                .catch((error) => {
+                    console.log(error);
+                });
+        }
     }
 }
 
@@ -170,14 +178,16 @@ export async function fetchLogin(data, navigate, setUserState, setError) {
         data,
     })
         .then((response) => {
+            console.log(response);
             navigate('/');
+            token = response.data.data.tokens.accessToken;
             const newData = { ...response.data.data, expire: Date.now() + 600000 };
-            console.log(newData);
             setUserState(newData);
             localStorage.setItem('user', JSON.stringify(newData));
         })
         .catch((error) => {
             setError('password', { message: error.response.data.message });
+            console.log(error);
         });
 }
 
@@ -196,11 +206,13 @@ export async function fetchSignup(data, navigate, setError) {
         },
     })
         .then((response) => {
+            console.log(response);
             navigate('/');
             alert('Admin 가입을 환영합니다!');
             console.log(response);
         })
         .catch((error) => {
             setError('password2', { message: error.response.data.message });
+            console.log(error);
         });
 }
